requirements.txt
# requirements.txt
# Core ML
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning (LSTM)
tensorflow>=2.15.0,<2.17.0

# AWS
boto3>=1.34.0
sagemaker>=2.200.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Development
jupyter>=1.0.0
pytest>=7.4.0

# Drift Detection
scipy>=1.11.0
